backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-multipart==0.0.9
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.3.0
pytesseract==0.3.10
